+++ experiments/ssg_test/lib/app.dart
@@ -1,14 +1,16 @@
-import 'package:jaspr/jaspr.dart';
+import 'package:jaspr/html.dart';
 
 part 'app.g.dart';
 
 @client
 class App extends StatelessComponent with _$App {
+  const App(this.title);
+
+  final String title;
+
   @override
   Iterable<Component> build(BuildContext context) sync* {
-    yield DomComponent(
-      tag: 'p',
-      child: Text('Hello World'),
-    );
+    yield h1([text(title)]);
+    yield p([text('Hello World')]);
   }
 }