+++ packages/jaspr_router/lib/src/registry/registry_stub.dart
@@ -1,0 +1,7 @@
+import '../route.dart';
+import 'registry.dart';
+
+class RouteRegistryImpl implements RouteRegistry {
+  @override
+  void registerRoutes(List<RouteBase> routes) {}
+}